Add: アイテム削除および更新時にユーザーのメールアドレスを確認するロジックを追加

+++ src/app/api/item/delete/[id]/route.js
@@ -6,9 +6,14 @@ export async function DELETE(request, context) {
     const reqBody = await request.json();
     try {
         await connectDB();
-        const { id } = await context.params;
-        await ItemModel.deleteOne({ _id: id }, reqBody);
-        return NextResponse.json({message: "delete success"});
+        const singleItem = await ItemModel.findById(context.params.id);
+        if (singleItem.email === reqBody.email) {
+            const { id } = await context.params;
+            await ItemModel.deleteOne({ _id: id }, reqBody);
+            return NextResponse.json({message: "delete success"});
+        } else {
+            return NextResponse.json({message: "delete failed"});
+        }
     } catch {
         return NextResponse.json({message: "delete failed"});
     }
